refactor(core): simplify skill attachment admission
