+++ packages/core/src/session.ts
@@ -144,13 +144,6 @@ type PendingInputRef = { readonly sessionID: SessionSchema.ID; readonly inputID:
 export class SkillNotFoundError extends Schema.TaggedErrorClass<SkillNotFoundError>()("Session.SkillNotFoundError", {
   skill: Skill.ID,
 }) {}
-export class SkillAttachmentError extends Schema.TaggedErrorClass<SkillAttachmentError>()(
-  "Session.SkillAttachmentError",
-  {
-    skill: Skill.ID,
-    message: Schema.String,
-  },
-) {}
 
 export class DestinationNotFoundError extends Schema.TaggedErrorClass<DestinationNotFoundError>()(
   "Session.DestinationNotFoundError",
@@ -229,10 +222,7 @@ export interface Interface {
     metadata?: Record<string, unknown>
     delivery?: SessionPending.Delivery
     resume?: boolean
-  }) => Effect.Effect<
-    SessionPending.User,
-    NotFoundError | PromptConflictError | AttachmentError | SkillNotFoundError | SkillAttachmentError
-  >
+  }) => Effect.Effect<SessionPending.User, NotFoundError | PromptConflictError | AttachmentError | SkillNotFoundError>
   /** Generates text from current Session context without admitting input or mutating history. */
   readonly generate: (input: {
     sessionID: SessionSchema.ID
@@ -256,7 +246,6 @@ export interface Interface {
     | PromptConflictError
     | AttachmentError
     | SkillNotFoundError
-    | SkillAttachmentError
     | Command.NotFoundError
     | Command.EvaluationError
   >
@@ -584,32 +573,16 @@ const layer = Layer.effect(
             // image attachment actually needs the resizer.
             const image = Image.Service.pipe(Effect.provide(locations.get(session.location)))
             const skills = Skill.Service.pipe(Effect.provide(locations.get(session.location)))
-            const messageID = input.id ?? SessionMessage.ID.create()
-            const delivery = input.delivery ?? "steer"
-            const previous =
-              input.id && input.skills?.length
-                ? yield* SessionPending.existing(db, {
-                    id: messageID,
-                    sessionID: input.sessionID,
-                    delivery,
-                  }).pipe(
-                    Effect.catchDefect((defect) =>
-                      defect instanceof SessionPending.LifecycleConflict
-                        ? new PromptConflictError({ sessionID: input.sessionID, messageID })
-                        : Effect.die(defect),
-                    ),
-                  )
-                : undefined
             const prompt = yield* resolvePrompt(
               { text: input.text, files: input.files, agents: input.agents, skills: input.skills },
               image,
               skills,
-              previous?.type === "user" ? previous.data.skills : undefined,
             ).pipe(Effect.provideService(FSUtil.Service, fs))
+            const messageID = input.id ?? SessionMessage.ID.create()
             const admittedInput = SessionPending.Message.make({
               type: "user",
               data: { ...prompt, metadata: input.metadata },
-              delivery,
+              delivery: input.delivery ?? "steer",
             })
             const admitted = yield* SessionPending.admit(db, bus, {
               id: messageID,
@@ -933,38 +906,25 @@ const resolvePrompt = Effect.fn("Session.resolvePrompt")(function* (
   input: PromptInput.Prompt,
   image: Effect.Effect<Image.Interface>,
   skills: Effect.Effect<Skill.Interface>,
-  previous: Prompt["skills"],
 ) {
   const fs = yield* FSUtil.Service
   const files = input.files
     ? yield* Effect.forEach(input.files, (file) => materializeAttachment(fs, file, image), { concurrency: 8 })
     : undefined
   const requested = input.skills
-  const reusable =
-    requested &&
-    previous &&
-    JSON.stringify(requested) === JSON.stringify(previous.map((skill) => ({ id: skill.id, mention: skill.mention })))
-      ? previous
-      : undefined
   const selected = yield* Effect.gen(function* () {
     if (!requested?.length) return undefined
-    if (reusable) return reusable
     const available = yield* (yield* skills).list()
-    return yield* Effect.forEach(requested, (attachment) =>
-      Effect.gen(function* () {
-        const skill = available.find((item) => item.id === attachment.id)
-        if (!skill) return yield* new SkillNotFoundError({ skill: attachment.id })
-        const output = yield* Skill.modelOutput(fs, skill).pipe(
-          Effect.mapError((error) => new SkillAttachmentError({ skill: skill.id, message: String(error) })),
-        )
-        return {
-          id: skill.id,
-          name: skill.name,
-          text: output.output,
-          mention: attachment.mention,
-        }
-      }),
-    )
+    return yield* Effect.forEach(requested, (attachment) => {
+      const skill = available.find((item) => item.id === attachment.id)
+      if (!skill) return Effect.fail(new SkillNotFoundError({ skill: attachment.id }))
+      return Effect.succeed({
+        id: skill.id,
+        name: skill.name,
+        text: Skill.toModelOutput(skill, []),
+        mention: attachment.mention,
+      })
+    })
   })
   return Prompt.make({ text: input.text, agents: input.agents, files, skills: selected?.length ? selected : undefined })
 })
