+++ packages/core/src/session/pending.ts
@@ -130,15 +130,6 @@ const promotedFromMessage = Effect.fn("SessionPending.promotedFromMessage")(func
   return yield* Effect.die(new LifecycleConflict({ id }))
 })
 
-export const existing = Effect.fn("SessionPending.existing")(function* (
-  db: DatabaseService,
-  input: PendingRef & { readonly delivery: Delivery },
-) {
-  const pending = yield* find(db, input.id)
-  if (pending !== undefined) return pending
-  return yield* promotedFromMessage(db, input.sessionID, input.id, input.delivery)
-})
-
 export const admit = Effect.fn("SessionPending.admit")(function* (
   db: DatabaseService,
   bus: Bus.Interface,
@@ -148,15 +139,13 @@ export const admit = Effect.fn("SessionPending.admit")(function* (
     readonly input: Message
   },
 ) {
-  const stored = yield* existing(db, {
-    id: request.id,
-    sessionID: request.sessionID,
-    delivery: request.input.delivery,
-  })
-  if (stored !== undefined) {
-    if (stored.type === "compaction") return yield* Effect.die(new LifecycleConflict({ id: request.id }))
-    return stored
+  const existing = yield* find(db, request.id)
+  if (existing !== undefined) {
+    if (existing.type === "compaction") return yield* Effect.die(new LifecycleConflict({ id: request.id }))
+    return existing
   }
+  const promoted = yield* promotedFromMessage(db, request.sessionID, request.id, request.input.delivery)
+  if (promoted !== undefined) return promoted
   return yield* bus
     .publish(SessionEvent.InputAdmitted, {
       inputID: request.id,
