+++ packages/core/src/skill.ts
@@ -38,8 +38,6 @@ export { Event } from "@opencode-ai/schema/skill"
 export const available = (skills: ReadonlyArray<Info>, agent: Agent.Info) =>
   skills.filter((skill) => Permission.evaluate("skill", skill.id, agent.permissions).effect !== "deny")
 
-const FILE_LIMIT = 10
-
 export const toModelOutput = (skill: Info, files: ReadonlyArray<string>) => {
   const directory = path.dirname(skill.location)
   return [
@@ -58,18 +56,6 @@ export const toModelOutput = (skill: Info, files: ReadonlyArray<string>) => {
     "</skill_content>",
   ].join("\n")
 }
-
-export const modelOutput = Effect.fn("Skill.modelOutput")(function* (fs: FSUtil.Interface, skill: Info) {
-  const directory = path.dirname(skill.location)
-  const files =
-    path.basename(skill.location) === "SKILL.md"
-      ? (yield* fs.scan("**/*", { cwd: directory, absolute: true, include: "file", dot: true }))
-          .filter((file) => path.basename(file) !== "SKILL.md")
-          .toSorted()
-          .slice(0, FILE_LIMIT)
-      : []
-  return { directory, output: toModelOutput(skill, files) }
-})
 
 const Frontmatter = Schema.Struct({
   name: Schema.String.pipe(Schema.optional),
